requirements.txt
asgiref==3.11.1
Django==6.0.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
pillow==12.1.1
sqlparse==0.5.5
